+++ Todolist.py
@@ -5,28 +5,53 @@ from pathlib import Path
 from audiorecorder import audiorecorder
 import streamlit as st
 import config
-import os 
+import os
+# import torch
+# from transformers import AutoModelForSpeechSeq2Seq, AutoProcessor, pipeline
+# import numpy as np
 
 os.environ["PATH"] += os.pathsep + r"C:ffmpeg\ffmpeg\bin"
-def ingest_query(query,model,sample_file):
-    response = model.generate_content([sample_file, query])
-    logger.info(response)
-    return response
+
+def generate_response(query,model,audio=None):
+    """
+    Uses gemini api to generate a response based on input. 
+
+    query: the query to the model based on the audio input.
+    model: Generative Model to be used for the task. default = genai.GenerativeModel('gemini-1.5-flash').
+    audio: task recorded as audio input for the breakdown, if audiorecoder is used. default = None.
+    """
+    if audio:
+        audio.export("audio/audio.wav", format="wav")
+        audio_query = genai.upload_file(path="audio/audio.wav", display_name="audio input query")
+        response = model.generate_content([audio_query, query])
+    else:
+        response = model.generate_content(query)
+    st.session_state['response'] = response
 
 if __name__ == "__main__":
+    if 'response' not in st.session_state:
+        st.session_state['response'] = None
+    st.header("AI-Powered To-Do List",  divider='rainbow')
+    st.subheader("Audio Recorder")
+
+
     genai.configure(api_key=config.google_api_key)
-    model = genai.GenerativeModel('gemini-1.5-flash')
+    gen_model = genai.GenerativeModel('gemini-1.5-flash')
     file_path = Path("audio").mkdir(parents=True, exist_ok=True)
-    st.title("Audio Recorder")
-    audio = audiorecorder("Click to record", "Click to stop recording")
-    if len(audio) > 0:
+    
+
+    audio_input_on = st.toggle("record voice through mic")
+    if audio_input_on:
+        query = "Listen to audio prompt and answer it in detail"
+        audio = audiorecorder("Click to record", "Click to stop recording")
         st.audio(audio.export().read())
-        audio.export("audio/audio.wav", format="wav")
-        # st.button("Record audio", type="primary", on_click=record_audio(file_path))
-        # input_file = Path("{file_path}/audio.wav")
-        # sample_file = genai.upload_file(path=input_file, display_name="Sample audio")
-        # user_audio_query = "Listen to the audio prompt, and answer it. The output response must be in markdown format"
-        # input = st.text_area(label="Input your query here.", placeholder="Listen to audio prompt and answer it")
-        # if input:
-        #     response = ingest_query(input, model, sameple_file)
-        #     st.markdown(response)
+        if audio:
+            if st.button("Submit Response", type="primary"):
+                generate_response(query=query, model=gen_model, audio=audio)
+    else:
+        input_query = st.text_area(label="Input your query here.", placeholder="I want to run a 10K in 1hour, currently I run a 10K in 70mins. Give me a breakdown on training regimen, i have 1 month to my race")
+        if st.button("Submit Response", type="primary"):
+            generate_response(query=input_query, model=gen_model)
+    
+    if st.session_state['response']:
+        st.markdown(st.session_state['response'].text)