+++ CHANGELOG.md
@@ -3,7 +3,20 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [Unreleased] - 2024-06-17
+
+Updates pushed in this commit - 
+
+### Added
+- Directly added both voice and text input 
+- These inputs are directly fed into the model to generate a response
  
+### Changed
+ 
+### Fixed
+
+
 ## [Unreleased] - 2024-06-11
  
 Here we write upgrades to the software as a whole.